LPD-101379 Migrate the fragment and collection constants to TypeScript
