+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/config/constants/textEditableTypes.ts
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {EDITABLE_TYPES} from './editableTypes';
+import {EDITABLE_TYPES, EditableType} from './editableTypes';
 
-export const TEXT_EDITABLE_TYPES = new Set([
+export const TEXT_EDITABLE_TYPES = new Set<EditableType>([
 	EDITABLE_TYPES.link,
 	EDITABLE_TYPES['rich-text'],
 	EDITABLE_TYPES.text,
