+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/config/constants/translationStatusType.ts
@@ -3,7 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export const FRAGMENT_CONFIGURATION_ROLES = {
-	advanced: 'advanced',
-	style: 'style',
-};
+export const TRANSLATION_STATUS_TYPE = {
+	default: 'default',
+	translated: 'translated',
+	translating: 'translating',
+	untranslated: 'untranslated',
+} as const;
